Rename `data_classes.dart` files
