+++ lib/flutter_widget_from_html.dart
@@ -1,5 +1,5 @@
 export 'src/builder.dart';
-export 'src/data_classes.dart';
+export 'src/data.dart';
 export 'src/helpers.dart';
 export 'src/html_widget.dart';
 export 'src/widget_factory.dart';
