+++ lib/src/data.dart
@@ -1,0 +1,1 @@
+export 'package:flutter_widget_from_html_core/src/core_data.dart';
