--- lib/src/data_classes.dart
@@ -1,1 +1,0 @@
-export 'package:flutter_widget_from_html_core/src/data_classes.dart';
