+++ packages/core/lib/flutter_widget_from_html_core.dart
@@ -1,5 +1,5 @@
 export 'src/builder.dart';
+export 'src/core_data.dart';
 export 'src/core_helpers.dart';
 export 'src/core_html_widget.dart';
 export 'src/core_widget_factory.dart';
-export 'src/data_classes.dart';
