+++ packages/core/lib/src/builder.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/widgets.dart';
 import 'package:html/dom.dart' as dom;
 
+import 'core_data.dart';
 import 'core_helpers.dart';
 import 'core_widget_factory.dart';
-import 'data_classes.dart';
 
 final _attrStyleRegExp = RegExp(r'([a-zA-Z\-]+)\s*:\s*([^;]*)');
