+++ packages/core/lib/src/core_data.dart
