+++ packages/core/lib/src/core_widget_factory.dart
@@ -5,9 +5,9 @@ import 'package:flutter/widgets.dart';
 import 'package:html/dom.dart' as dom;
 
 import 'builder.dart';
+import 'core_data.dart';
 import 'core_helpers.dart';
 import 'core_html_widget.dart';
-import 'data_classes.dart';
 
 part 'ops/style_bg_color.dart';
 part 'ops/style_margin.dart';
